[Hotfix] Fixed Infinite Registration
The reason for the infinite registration bug existing is because of a
simple error in the order of asynchronous operations. Allow me to
elaborate. This was the code I had in auth.dart:

`await user.sendEmailVerification()
await Database(uid: user.uid).updateUserData(authUser);`

Most people would think that registering the user in the database would
take precedence over sending verification. Well, not me! I decided to
write it as in the example shown, leading to unnecessary delay for the
database writing operation. This eventually bit me in the back as I had
this code in a method I used to get the user in database.dart:

`return getUserBySnapshot(await
_userCollectionReference.document(uid).get())`

And I call this method once the user clicks the register button. But
when he clicks said button, he is also being registered at the same
time. So that email verification managed to slow down the creation of
the user so much that I was reading from the database before even having
written to it! Lesson learned!