[Hotfix] Fixed external Auth app detach bug
Fixed a bug which occured upon a detach of the current Activity (the app
is put in the background) and a re-entry into the app. An exception was
thrown, but the issue was resolved by omitting the external Auth
register page from the stack (through Navigator.pop(context)) and checking whether external auth widget properties themselves were valid in the following re-entry of the app.