[Update] Added Notification skeleton (NOT WORKING)
Added a notification skeleton and config for Android. iOS is, again, not
supported. Begain work on notitifactions with the creation of several
classes like Messaging and NotificationList to aid in the main segments
of the feature. Unfortunately, several changes were made to
database.dart which broke major parts of the auth and user display. DO NOT ALLOW this version to go
through at any cost. Will be fixed very, very soon in the next commit
(it's being committed now just because the dev wants to get some
freaking sleep after trying to fix it for hours on end).